+++ reproc++/src/reproc.cpp
@@ -59,12 +59,10 @@ process::~process() noexcept
 process::process(process &&) noexcept = default; // NOLINT
 process &process::operator=(process &&) noexcept = default;
 
-std::error_code process::start(int argc,
-                               const char *const *argv,
+std::error_code process::start(const char *const *argv,
                                const char *working_directory) noexcept
 {
-  REPROC_ERROR error = reproc_start(process_.get(), argc, argv,
-                                    working_directory);
+  REPROC_ERROR error = reproc_start(process_.get(), argv, working_directory);
   return error_to_error_code(error);
 }