+++ reproc/examples/git-status.c
@@ -19,11 +19,8 @@ int main(void)
   // `reproc_t` stores necessary information between calls to reproc's API.
   reproc_t git_status;
 
-  /* `reproc_start` imposes the same restrictions on `argc` and `argv` as the
-  regular main function of C and C++ applications. `argv` is required to end
-  with a `NULL` element and `argc` should be one less than the total amount of
-  elements in `argv` (it doesn't count the `NULL`). */
-  int argc = 2;
+  /* `argv` must start with the name of the program to execute and must end with
+  a `NULL` value. */
   const char *argv[3] = { "git", "status", NULL };
 
   // Most of reproc's API functions return a value from `REPROC_ERROR` to
@@ -31,10 +28,10 @@ int main(void)
   // returned.
   REPROC_ERROR error = REPROC_SUCCESS;
 
-  /* `reproc_start` takes a child process instance (`reproc_t`), argc, argv
+  /* `reproc_start` takes a child process instance (`reproc_t`), argv
   and the working directory of the child process. If the working directory is
   `NULL` the working directory of the parent process is used. */
-  error = reproc_start(&git_status, argc, argv, NULL);
+  error = reproc_start(&git_status, argv, NULL);
 
   /* reproc exposes a single error enum `REPROC_ERROR` which contains errors
   specific to reproc and `REPROC_ERROR_SYSTEM` to indicate a system error