+++ reproc/src/posix/reproc.c
@@ -11,19 +11,12 @@
 #include <unistd.h>
 
 REPROC_ERROR reproc_start(reproc_t *process,
-                          int argc,
                           const char *const *argv,
                           const char *working_directory)
 {
   assert(process);
-
-  assert(argc > 0);
   assert(argv);
-  assert(argv[argc] == NULL);
-
-  for (int i = 0; i < argc; i++) {
-    assert(argv[i]);
-  }
+  assert(argv[0] != NULL);
 
   process->running = false;