+++ reproc/test/argv.cpp
@@ -11,12 +11,10 @@ TEST_SUITE("reproc")
   {
     reproc_t process;
 
-    static constexpr unsigned int ARGV_SIZE = 4;
     std::array<const char *, 4> argv = { "reproc/resources/argv", "argument 1",
                                          "argument 2", nullptr };
 
-    REPROC_ERROR error = reproc_start(&process, ARGV_SIZE - 1, argv.data(),
-                                      nullptr);
+    REPROC_ERROR error = reproc_start(&process, argv.data(), nullptr);
     REQUIRE(!error);
 
     std::string output;